+++ src/lib/zones/mod.rs
@@ -1,4 +1,5 @@
+pub mod statistics;
 pub mod skeleton;
 pub mod virtual_line;
 pub mod zones;
-pub use self::{skeleton::*, virtual_line::*, zones::*, zones::geometry::*, zones::geojson::*};
+pub use self::{statistics::*, skeleton::*, virtual_line::*, zones::*, zones::geometry::*, zones::geojson::*};
