+++ src/lib/zones/zones.rs
@@ -8,7 +8,7 @@ use std::collections::hash_map::Entry::{
     Vacant
 };
 use uuid::Uuid;
-use chrono::{DateTime, TimeZone, Utc};
+use chrono::{DateTime, Utc};
 
 use geometry::PointsOrientation;
 use geometry::{
@@ -38,39 +38,8 @@ use crate::lib::spatial::SpatialConverter;
 use crate::lib::spatial::epsg::lonlat_to_meters;
 use crate::lib::spatial::compute_center;
 use crate::lib::spatial::haversine;
-use crate::lib::zones::{Skeleton, VirtualLine};
+use crate::lib::zones::{VehicleTypeParameters, Statistics, Skeleton, VirtualLine};
 
-#[derive(Debug)]
-pub struct Statistics {
-    pub period_start: DateTime<Utc>,
-    pub period_end: DateTime<Utc>,
-    pub vehicles_data: HashMap<String, VehicleTypeParameters>
-}
-
-impl Statistics {
-    pub fn default() -> Self {
-        Statistics{
-            period_start: TimeZone::with_ymd_and_hms(&Utc, 1970, 1, 1, 0, 0, 0).unwrap(),
-            period_end: TimeZone::with_ymd_and_hms(&Utc, 1970, 1, 1, 0, 0, 0).unwrap(),
-            vehicles_data: HashMap::new()
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct VehicleTypeParameters {
-    pub avg_speed: f32,
-    pub sum_intensity: u32
-}
-
-impl VehicleTypeParameters {
-    pub fn default() -> Self {
-        VehicleTypeParameters{
-            avg_speed: -1.0,
-            sum_intensity: 0,
-        }
-    }
-}
 
 #[derive(Debug)]
 struct ObjectInfo {
